1-listint_len.c

+++ 0x13-more_singly_linked_lists/1-listint_len.c
@@ -2,18 +2,19 @@
 #include "lists.h"
 
 /**
- * h - pointer to struct
+ * listint_len - list nodes
+ * @h - pointer to struct
  *
+ * Return: list nodes
  */
-
 size_t listint_len(const listint_t *h)
 {
-        int x = 0;
+	size_t x = 0;
 
-        while (h != NULL)
-        {
-                x++;
-                h = h -> next;
-        }
-        return (x);
+	while (h != NULL)
+	{
+		x++;
+		h = h -> next;
+	}
+	return (x);
 }
